+++ scripts/metadata_comparison/metadata_comparison/extractor.py
@@ -24,7 +24,7 @@ from google.cloud import storage
 import google.auth
 from googleapiclient.discovery import build as google_client_build
 import logging as log
-from metadata_comparison.lib_argument_regex import *
+from metadata_comparison.lib.argument_regex import *
 
 logger = log.getLogger('metadata_comparison.extractor')
