+++ scripts/metadata_comparison/metadata_comparison/lib/argument_regex.py
